--- packages/front-end/src/app/components/EGRunPipelineStepper.vue
@@ -1,246 +1,0 @@
-<script setup lang="ts">
-  import { useRunStore } from '@FE/stores';
-  import { Pipeline as SeqeraPipeline } from '@easy-genomics/shared-lib/src/app/types/nf-tower/nextflow-tower-api';
-
-  const props = defineProps<{
-    schema: object;
-    params: object;
-    pipelineId: string;
-    labName: string;
-  }>();
-
-  const $route = useRoute();
-  const runStore = useRunStore();
-  const seqeraPipelineStore = useSeqeraPipelinesStore();
-
-  const seqeraRunTempId = $route.query.seqeraRunTempId as string;
-
-  const wipSeqeraRun = computed<WipSeqeraRunData | undefined>(() => runStore.wipSeqeraRuns[seqeraRunTempId]);
-
-  const pipeline = computed<SeqeraPipeline | undefined>(() => seqeraPipelineStore.pipelines[props.pipelineId]);
-
-  const labId = $route.params.labId as string;
-
-  const selectedIndex = ref(0);
-  const hasLaunched = ref(false);
-  const resetKey = ref(0);
-
-  const emit = defineEmits(['has-launched', 'reset-run-pipeline']);
-
-  const items = ref([
-    {
-      disabled: false,
-      key: 'details',
-      label: 'Run Details',
-    },
-    {
-      disabled: true,
-      key: 'upload',
-      label: 'Upload Data',
-    },
-    {
-      disabled: true,
-      key: 'parameters',
-      label: 'Edit Parameters',
-    },
-    {
-      disabled: true,
-      key: 'review',
-      label: 'Review Pipeline',
-    },
-  ]);
-
-  // Note: the UTabs :ui attribute has to be defined locally in this file - if it is imported from another file,
-  //  Tailwind won't pick up and include the classes used and styles will be missing.
-  // To keep the tab styling consistent throughout the app, any changes made here need to be duplicated to all other
-  //  UTabs that use an "EGTabsStyles" as input to the :ui attribute.
-  const EGTabsStyles = {
-    base: 'focus:outline-none',
-    list: {
-      base: '!flex rounded-none mb-6 mt-0',
-      padding: 'p-0',
-      height: 'h-14',
-      marker: {
-        background: '',
-        shadow: '',
-      },
-      tab: {
-        base: 'font-serif w-auto mr-3 rounded-xl border border-solid',
-        background: '',
-
-        // TODO: this is a hotfix for what seems to be a bug with UTabs where it reports the wrong tab as selected
-        // we have to handle the styling manually ourselves
-        // active: 'text-white bg-primary border-primary',
-        active: '',
-
-        inactive: 'font-serif text-text-body border-background-dark-grey',
-        height: '',
-        padding: 'px-5 py-2',
-        size: 'text-sm',
-      },
-    },
-  };
-
-  /**
-   * Set the enabled state of a step in the stepper
-   * @param step
-   * @param isEnabled
-   */
-  function setStepEnabled(step: string, isEnabled: boolean) {
-    const stepIndex = items.value.findIndex((item) => item.key === step);
-    if (stepIndex > -1) {
-      if (isEnabled) {
-        items.value[stepIndex].disabled = false;
-      } else {
-        // If the step is disabled, disable all subsequent steps
-        disableStepsFrom(stepIndex);
-      }
-    }
-  }
-
-  /**
-   * Disable all steps from the given index
-   * @param index
-   */
-  function disableStepsFrom(index: number) {
-    for (let i = index; i < items.value.length; i++) {
-      items.value[i].disabled = true;
-    }
-  }
-
-  function nextStep(val: string) {
-    setStepEnabled(val, true);
-    selectedIndex.value = clampIndex(selectedIndex.value + 1);
-  }
-
-  function clampIndex(index: number) {
-    return Math.min(items.value.length - 1, Math.max(0, index));
-  }
-
-  function previousStep() {
-    selectedIndex.value = clampIndex(selectedIndex.value - 1);
-  }
-
-  function disableAllSteps() {
-    items.value.forEach((item) => {
-      item.disabled = true;
-    });
-  }
-
-  function enableAllSteps() {
-    items.value.forEach((item) => {
-      item.disabled = false;
-    });
-  }
-
-  function handleLaunchSuccess() {
-    hasLaunched.value = true;
-    emit('has-launched');
-  }
-
-  function handleSubmitLaunchRequest() {
-    disableAllSteps();
-  }
-
-  function handleSubmitLaunchRequestError() {
-    enableAllSteps();
-  }
-</script>
-
-<template>
-  <div :key="resetKey">
-    <UTabs :items="items" :ui="EGTabsStyles" class="UTabs" v-model="selectedIndex">
-      <template #default="{ item, index, selected }">
-        <div
-          class="relative flex items-center gap-2 truncate"
-          :class="selectedIndex === index && !hasLaunched ? 'active-tab' : ''"
-        >
-          <UIcon
-            v-if="selectedIndex > index || hasLaunched"
-            name="i-heroicons-check-20-solid"
-            class="text-primary h-4 w-4 flex-shrink-0"
-          />
-          <span :class="selectedIndex > index || hasLaunched ? 'text-primary' : ''">{{ item.label }}</span>
-          <span v-if="selected" class="bg-primary-500 dark:bg-primary-400 absolute -right-4 h-2 w-2 rounded-full" />
-        </div>
-      </template>
-
-      <template #item="{ item, index }">
-        <div v-if="!hasLaunched">
-          <!-- Run Details -->
-          <template v-if="items[selectedIndex].key === 'details'">
-            <EGRunFormRunDetails
-              pipeline-or-workflow="Pipeline"
-              :pipeline-or-workflow-name="pipeline?.name"
-              :initial-run-name="wipSeqeraRun?.runName || ''"
-              :pipeline-or-workflow-description="pipeline?.description"
-              :wip-run-update-function="runStore.updateWipSeqeraRun"
-              :wip-run-temp-id="seqeraRunTempId"
-              @next-step="() => nextStep('upload')"
-              @step-validated="setStepEnabled('upload', $event)"
-            />
-          </template>
-
-          <!-- Upload Data -->
-          <template v-if="items[selectedIndex].key === 'upload'">
-            <EGRunFormUploadData
-              :lab-id="labId"
-              :lab-name="labName"
-              :pipeline-or-workflow-name="pipeline.name"
-              :wip-run="wipSeqeraRun"
-              :wip-run-update-function="runStore.updateWipSeqeraRun"
-              :wip-run-temp-id="seqeraRunTempId"
-              @next-step="() => nextStep('parameters')"
-              @previous-step="previousStep"
-              @step-validated="setStepEnabled('parameters', $event)"
-            />
-          </template>
-
-          <!-- Edit Parameters -->
-          <template v-if="items[selectedIndex].key === 'parameters'">
-            <EGRunPipelineFormEditParameters
-              :params="params"
-              :schema="schema"
-              :seqera-run-temp-id="seqeraRunTempId"
-              @next-step="() => nextStep('review')"
-              @previous-step="previousStep"
-            />
-          </template>
-
-          <!-- Review Pipeline -->
-          <template v-if="items[selectedIndex].key === 'review'">
-            <EGRunPipelineFormReview
-              :schema="props.schema"
-              :params="wipSeqeraRun?.params"
-              :pipeline-id="pipelineId"
-              @submit-launch-request="handleSubmitLaunchRequest"
-              @submit-launch-request-error="handleSubmitLaunchRequestError"
-              @has-launched="handleLaunchSuccess"
-              @previous-tab="previousStep"
-            />
-          </template>
-        </div>
-      </template>
-    </UTabs>
-
-    <template v-if="hasLaunched">
-      <EGEmptyDataCTA
-        message="Your Workflow Run has Launched! Check on your progress via Runs."
-        :primary-button-action="() => $router.push(`/labs/${labId}?tab=Lab+Runs`)"
-        primary-button-label="Back to Runs"
-        :secondary-button-action="() => emit('reset-run-pipeline')"
-        secondary-button-label="Launch Another Workflow Run"
-        img-src="/images/empty-state-launched.jpg"
-      />
-    </template>
-  </div>
-</template>
-
-<style scoped lang="scss">
-  // this styling should be handled by the UTabs :ui, but isn't working correctly - see the EGTabsStyles for info
-  :deep(button:has(> .active-tab)) {
-    color: white;
-    background: #5524e0;
-    border: solid 1px #5524e0;
-  }
-</style>
